requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.0.0
python-docx>=0.8.11
tenacity>=8.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
PyPDF2>=3.0.0
Authlib>=1.3.0
requests>=2.31.0